refactor download status API to live in InstrumentChannel, and separate UI concerns from data reporting concerns

+++ scopehal/TestWaveformSource.h
@@ -62,7 +62,7 @@ public:
 		float period,
 		int64_t sampleperiod,
 		size_t depth,
-		std::function<void(int)> downloadCallback,
+		std::function<void(float)> downloadCallback,
 		float noise_stdev = 0.01);
 
 	WaveformBase* GenerateNoisySinewaveSum(
@@ -73,7 +73,7 @@ public:
 		float period2,
 		int64_t sampleperiod,
 		size_t depth,
-		std::function<void(int)> downloadCallback,
+		std::function<void(float)> downloadCallback,
 		float noise_stdev = 0.01);
 
 	WaveformBase* GeneratePRBS31(
@@ -83,7 +83,7 @@ public:
 		float period,
 		int64_t sampleperiod,
 		size_t depth,
-		std::function<void(int)> downloadCallback,
+		std::function<void(float)> downloadCallback,
 		bool lpf = true,
 		float noise_stdev = 0.01);
 
@@ -94,7 +94,7 @@ public:
 		float period,
 		int64_t sampleperiod,
 		size_t depth,
-		std::function<void(int)> downloadCallback,
+		std::function<void(float)> downloadCallback,
 		bool lpf = true,
 		float noise_stdev = 0.01);
